+++ packages/french-verbs-lefff/example.js
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
 const LefffVerbs = require('./dist/conjugations.json');
 
 // indicatif présent
