+++ packages/rosaenlg-pug-code-gen/src/LanguageCodeGenFrench.ts
@@ -6,8 +6,10 @@
 
 import { LanguageCodeGen, VerbInfo, WordInfo, AdjectiveInfo } from './LanguageCodeGen';
 import { getWordInfo } from 'french-words';
-import frenchVerbsDict from 'french-verbs-lefff';
+import frenchVerbsDict from 'french-verbs-lefff/dist/conjugations.json';
+import { VerbsInfo } from 'french-verbs-lefff';
 import frenchWordsGenderLefff from 'french-words-gender-lefff';
+import { GenderList } from 'french-words';
 import { getAdjectiveInfo } from 'french-adjectives-wrapper';
 import { getVerbInfo } from 'french-verbs';
 
@@ -18,10 +20,10 @@ export class LanguageCodeGenFrench extends LanguageCodeGen {
   hasFlexAdjectives = true;
 
   getVerbInfo(verb: string): VerbInfo {
-    return getVerbInfo(frenchVerbsDict, verb);
+    return getVerbInfo(frenchVerbsDict as VerbsInfo, verb);
   }
   getWordInfo(word: string): WordInfo {
-    return getWordInfo(frenchWordsGenderLefff, word);
+    return getWordInfo(frenchWordsGenderLefff as GenderList, word);
   }
   getAdjectiveInfo(adjective: string): AdjectiveInfo {
     // NB no need to give an custom list here
